avoid unnecessary float/double conversions

+++ src/main/java/com/github/stephengold/joltjni/FixedConstraintSettings.java
@@ -79,9 +79,9 @@ public class FixedConstraintSettings extends TwoBodyConstraintSettings {
      */
     public Vec3 getAxisX1() {
         long settingsVa = va();
-        double x = getAxisX1X(settingsVa);
-        double y = getAxisX1Y(settingsVa);
-        double z = getAxisX1Z(settingsVa);
+        float x = getAxisX1X(settingsVa);
+        float y = getAxisX1Y(settingsVa);
+        float z = getAxisX1Z(settingsVa);
         Vec3 result = new Vec3(x, y, z);
 
         return result;
@@ -95,9 +95,9 @@ public class FixedConstraintSettings extends TwoBodyConstraintSettings {
      */
     public Vec3 getAxisX2() {
         long settingsVa = va();
-        double x = getAxisX2X(settingsVa);
-        double y = getAxisX2Y(settingsVa);
-        double z = getAxisX2Z(settingsVa);
+        float x = getAxisX2X(settingsVa);
+        float y = getAxisX2Y(settingsVa);
+        float z = getAxisX2Z(settingsVa);
         Vec3 result = new Vec3(x, y, z);
 
         return result;
@@ -111,9 +111,9 @@ public class FixedConstraintSettings extends TwoBodyConstraintSettings {
      */
     public Vec3 getAxisY1() {
         long settingsVa = va();
-        double x = getAxisY1X(settingsVa);
-        double y = getAxisY1Y(settingsVa);
-        double z = getAxisY1Z(settingsVa);
+        float x = getAxisY1X(settingsVa);
+        float y = getAxisY1Y(settingsVa);
+        float z = getAxisY1Z(settingsVa);
         Vec3 result = new Vec3(x, y, z);
 
         return result;
@@ -127,9 +127,9 @@ public class FixedConstraintSettings extends TwoBodyConstraintSettings {
      */
     public Vec3 getAxisY2() {
         long settingsVa = va();
-        double x = getAxisY2X(settingsVa);
-        double y = getAxisY2Y(settingsVa);
-        double z = getAxisY2Z(settingsVa);
+        float x = getAxisY2X(settingsVa);
+        float y = getAxisY2Y(settingsVa);
+        float z = getAxisY2Z(settingsVa);
         Vec3 result = new Vec3(x, y, z);
 
         return result;
